pyproject.toml
[tool.poetry]
name = "oz-mini-project"
version = "0.1.0"
description = ""
authors = ["hwangtate <ghkdxodud10253504@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.5"
django = "5.1.1"
djangorestframework = "3.15.2"
python-dotenv = "1.0.1"
psycopg2 = "2.9.9"
drf-spectacular = "0.27.2"
djangorestframework-simplejwt = "^5.3.1"

[tool.poetry.group.dev.dependencies]
black = "24.8.0"
isort = "5.13.2"
mypy = "1.11.2"
coverage = "7.6.1"
django-stubs = "5.0.4"
djangorestframework-stubs = "^3.15.1"

[tool.black]
line-length = 150

[tool.isort]
profile = "black"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = 3.12
strict = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[tool.coverage.report]
fail_under = 70

[tool.coverage.run]
omit = [
  ".mypy_cache/**/*",
  "*/migrations/*",
  "*/*test*.py",
]

[tool.django-stubs]
django_settings_module = "miniproject.settings"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
